+++ Plugins/Experimental/Experimental.hpp
@@ -6,6 +6,7 @@ namespace Experimental {
 
 class SuppressPlayerLoginInfo;
 class AdjustReputationFix;
+class DisableLevelUpValidation;
 
 class Experimental : public NWNXLib::Plugin
 {
@@ -16,6 +17,7 @@ public:
 private:
     std::unique_ptr<SuppressPlayerLoginInfo> m_SuppressPlayerLoginInfo;
     std::unique_ptr<AdjustReputationFix> m_AdjustReputationFix;
+    std::unique_ptr<DisableLevelUpValidation> m_DisableLevelUpValidation;
 };
 
 }