+++ CHANGELOG.md
@@ -8,13 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 https://github.com/nwnxee/unified/compare/build8193.20...HEAD
 
 ### Added
-- N/A
+- Experimental: Added `NWNX_EXPERIMENTAL_DISABLE_LEVELUP_VALIDATION` to disable levelup validation.
 
 ##### New Plugins
 - N/A
 
 ##### New NWScript Functions
-- N/A
+- Util: GetInstructionLimit()
+- Util: {Get|Set}InstructionsExecuted();
 
 ### Changed
 - N/A
@@ -29,7 +30,7 @@ https://github.com/nwnxee/unified/compare/build8193.20...HEAD
 - N/A
 
 ## 8193.16
-https://github.com/nwnxee/unified/compare/build8193.16...build8193.19
+https://github.com/nwnxee/unified/compare/build8193.16...build8193.20
 
 ### Added
 - Chat: targeted messages can now be sent on the Party or DM channels